Refactor certificate expiry handling and notifications
Updated certificate expiry date handling and notification logic.

The previous certificate expiry date logic retrieves the expiry date as UTC not local.

Enhanced the code so that to eliminate unnecessary overwhelming email notifications, for every change in the certificate. It is recommended to be disabled so that to focust on the expiring and expired certificates.

Enhanced the code so that to eliminate unnecessary overwhelming email notifications, for every change search string. It is recommended to be disabled so that to focus on the notifications that require action.